# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/Auto.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/AutoPaths.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/subsystems/Intake.java
@@ -33,8 +33,7 @@ public class Intake {
     public Command setIn = instant(() -> intake.set(1));
     public Command setOut = instant(() -> intake.set(-1));
     public Command turnOff = instant(() -> intake.set(0));
-    public Command run(int power){
+    public Command run(double power){
         return instant(() -> intake.set(power));
     }
-
 }
